refactor(bpane-host): extract tile capture thread into tile_loop module
Move the 1550-line spawn_blocking closure into a TileCaptureThread struct
with structured state management:

- tile_loop/mod.rs (251 lines): struct definition with ~40 fields,
  new() constructor that initialises X11 capture, grid, and damage tracker
- tile_loop/run.rs (1478 lines): run() method containing the per-frame
  processing loop, with constants and frame logic

The struct eliminates ~10 _for_tile Arc clones and bundles all tile
capture state into named fields. main.rs reduced from 2616 to 1096 lines
(-1520). All 287 tests pass.

Note: run.rs exceeds the 200-line target — further decomposition into
sub-methods (cdp hints, scroll orchestration, classification, emission)
is tracked as a follow-up task.
