+++ code/apps/bpane-host/src/main.rs
@@ -14,6 +14,7 @@ mod region;
 mod resize;
 mod scroll;
 pub mod tiles;
+mod tile_loop;
 mod video_classify;
 mod video_region;
 
@@ -422,1571 +423,50 @@ async fn run_ffmpeg_session(
     let (cache_miss_tx, cache_miss_rx) = std::sync::mpsc::channel::<(u32, u16, u16, u64)>();
 
     // ── Tile capture thread ─────────────────────────────────────────
-    // Runs alongside FFmpeg: captures screen via X11 GetImage, hashes
-    // 64×64 tiles, emits Fill/QOI/CacheHit frames on the Tiles channel.
-    // This gives lossless tile delivery for static content while FFmpeg
-    // handles motion via H.264.
-    let display_for_tiles = display.to_string();
     let (tile_tx, mut tile_rx) = mpsc::channel::<Frame>(256);
-    let tile_w = args.width;
-    let tile_h = args.height;
-    let h264_mode_for_tile = h264_mode;
-    let cmd_tx_for_tile = cmd_tx.clone();
-    let tiles_active_for_tile = tiles_active.clone();
-    let video_tile_info_for_tile = video_tile_info.clone();
-    let browser_video_hint_for_tile = browser_video_hint.clone();
-    let input_activity_for_tile = input_activity_ms.clone();
-    let tile_size_for_tiles = tile_size_from_env();
-    let tile_codec_for_tiles = tile_codec_from_env();
-    let scroll_copy_quantum_px_for_tile = scroll_copy_quantum_px;
+    let display_for_tiles = display.to_string();
+    let tile_init_w = args.width;
+    let tile_init_h = args.height;
+    let cmd_tx_for_tiles = cmd_tx.clone();
+    let tiles_active_for_tiles = tiles_active.clone();
+    let video_tile_info_for_tiles = video_tile_info.clone();
+    let browser_video_hint_for_tiles = browser_video_hint.clone();
+    let input_activity_for_tiles = input_activity_ms.clone();
     let tile_thread = tokio::task::spawn_blocking(move || {
-        use capture::CaptureBackend;
-
-        let mut cap = match capture::x11::X11CaptureBackend::new(&display_for_tiles, tile_w, tile_h)
-        {
-            Ok(c) => c,
-            Err(e) => {
-                warn!("tile capture: X11 backend init failed: {e}");
-                return;
-            }
-        };
-
-        // Use actual screen resolution from X11, not args (screen may have resized).
-        let (init_w, init_h) = cap.resolution();
-        let mut screen_w = init_w as u16;
-        let mut screen_h = init_h as u16;
-        let tile_size: u16 = tile_size_for_tiles;
-        let tile_codec = tile_codec_for_tiles;
-        info!(
-            "tile capture: active ({}x{}, tile_size={}, codec={:?})",
-            screen_w, screen_h, tile_size, tile_codec
-        );
-
-        // ── Scroll displacement detection state ───────────────────
-        let mut prev_frame: Option<Vec<u8>> = None;
-        let mut content_origin_y: i64 = 0;
-        let mut grid_offset_y: u16 = 0;
-
-        let mut grid = tiles::TileGrid::new(screen_w, screen_h, tile_size);
-        let mut emitter = tiles::emitter::TileEmitter::with_codec(grid.cols, grid.rows, tile_codec);
-        let mut h264_enabled = h264_mode_for_tile.starts_enabled();
-        let mut region_committer = video_region::RegionCommitter::new();
-
-        // Per-tile tracking for robust video classification.
-        // Uses multi-signal scoring + hysteresis to avoid flip/flop.
-        let total_tiles = grid.cols as usize * grid.rows as usize;
-        let mut prev_hashes: Vec<u64> = vec![0; total_tiles];
-        let mut video_scores: Vec<i8> = vec![0; total_tiles];
-        let mut non_candidate_streaks: Vec<u8> = vec![0; total_tiles];
-        let mut video_hold_frames: Vec<u8> = vec![0; total_tiles];
-        let mut video_latched: Vec<bool> = vec![false; total_tiles];
-        let mut candidate_mask: Vec<bool> = vec![false; total_tiles];
-        let mut changed_mask: Vec<bool> = vec![false; total_tiles];
-        let mut text_like_mask: Vec<bool> = vec![false; total_tiles];
-        const VIDEO_ENTER_SCORE: i8 = 5;
-        const VIDEO_EXIT_SCORE: i8 = -1;
-        const VIDEO_MAX_SCORE: i8 = 20;
-        const VIDEO_DECAY_STREAK: u8 = 18;
-        const VIDEO_MIN_HOLD_FRAMES: u8 = 12;
-        const MIN_SCROLL_DY_PX: i32 = 2;
-        const INPUT_MIN_SCROLL_DY_PX: i32 = 4;
-        const MAX_CDP_SCROLL_DY_PX: i64 = scroll::CONTENT_SCROLL_SEARCH_MAX_PX as i64;
-        const CDP_CONTENT_DY_DIVERGENCE_LOG_PX: i32 = 3;
-        const INPUT_SCROLL_MIN_CONFIDENCE: f32 = 0.80;
-        const NO_INPUT_SCROLL_MIN_CONFIDENCE: f32 = 0.86;
-        const SCROLL_SUPPRESS_VIDEO_FRAMES: u8 = 14;
-        const SCROLL_RESIDUAL_FULL_REPAINT_RATIO: f32 = scroll::SCROLL_RESIDUAL_FULL_REPAINT_RATIO_DEFAULT;
-        // Reduce per-tick bandwidth on sub-tile scroll when residual diff is
-        // broad/noisy: send exposed strip while scrolling, then one repair.
-        const SCROLL_THIN_MODE_RESIDUAL_RATIO: f32 = 0.14;
-        const SCROLL_THIN_REPAIR_QUIET_FRAMES: u8 = 5;
-        let scroll_thin_mode_enabled = env_bool("BPANE_SCROLL_THIN_MODE", false);
-        const REGION_RECONFIG_STABLE_FRAMES: u8 = 2;
-        const REGION_RECONFIG_MIN_INTERVAL_MS: u64 = 350;
-        const REGION_MIN_CANDIDATES: u32 = 6;
-        const REGION_DENSE_CANDIDATES: u32 = 18;
-        const MIN_VIDEO_BBOX_WIDTH_TILES: u16 = 4;
-        const MIN_VIDEO_BBOX_HEIGHT_TILES: u16 = 3;
-        const MIN_VIDEO_BBOX_AREA_RATIO: f32 = 0.10;
-        // Video hints stay strict so tiny previews do not activate ROI H.264.
-        // Focused editable controls stay on the tile path and use QOI overrides.
-        let min_cdp_video_width_px =
-            env_u32_clamped("BPANE_CDP_MIN_VIDEO_WIDTH", 320, 2, 4096) & !1;
-        let min_cdp_video_height_px =
-            env_u32_clamped("BPANE_CDP_MIN_VIDEO_HEIGHT", 180, 2, 4096) & !1;
-        let min_cdp_video_area_ratio =
-            env_f32_clamped("BPANE_CDP_MIN_VIDEO_AREA_RATIO", 0.08, 0.01, 0.95);
-        const EDITABLE_QOI_TILE_MARGIN: u16 = 2;
-        const EDITABLE_HINT_HOLD_MS: u64 = 450;
-        const KEY_INPUT_QOI_BOOST_MS: u64 = 800;
-        const MIN_CHANGED_VIDEO_TILES_FOR_H264: u32 = 8;
-        let video_click_arm_ms =
-            env_u32_clamped("BPANE_VIDEO_CLICK_ARM_MS", 8_000, 250, 60_000) as u64;
-        const CLICK_LATCH_RESET_FRAMES: u8 = 20;
-        const H264_MIN_ON_DURATION_MS: u64 = 900;
-        let video_classification_enabled = !matches!(h264_mode_for_tile, H264Mode::Off);
-
-        // Bounding box stability: video regions have a stable spatial footprint
-        // while scroll shifts the changing region each frame. Track the bbox of
-        // video-like changing tiles and require temporal stability before
-        // allowing VideoMotion classification.
-        let mut prev_video_bbox: Option<(u16, u16, u16, u16)> = None; // (min_col, min_row, max_col, max_row)
-        let mut stable_bbox_frames: u8 = 0;
-        let mut scroll_cooldown_frames: u8 = 0;
-        let mut last_left_click: Option<(u16, u16, std::time::Instant)> = None;
-        let mut click_armed = video_region::ClickArmedState::new();
-        let mut editable_hint = video_region::EditableHintState::new();
-        // Scroll residual telemetry (host-side cumulative counters).
-        let mut scroll_residual_batches_total: u64 = 0;
-        let mut scroll_residual_fallback_full_total: u64 = 0;
-        let mut scroll_potential_tiles_total: u64 = 0;
-        let mut scroll_residual_tiles_total: u64 = 0;
-        let mut scroll_saved_tiles_total: u64 = 0;
-        let mut scroll_thin_mode_active: bool = false;
-        let mut scroll_residual_was_active: bool = false;
-        let mut scroll_quiet_frames: u8 = 0;
-        // Cached scroll region from last detected scroll frame — used to split
-        // static/content tiles during reconciliation and non-scroll frames.
-        let mut last_scroll_region_top: u16 = 0;
-        let mut last_scroll_region_bottom: u16 = screen_h;
-        let mut last_scroll_region_right: u16 = screen_w;
-        let mut scroll_thin_batches_total: u64 = 0;
-        let mut scroll_thin_repairs_total: u64 = 0;
-        let mut scroll_active_capture_frames_remaining: u8 = 0;
-        let mut last_cdp_hint_seq: u64 = 0;
-        let mut last_cdp_scroll_y: Option<i64> = None;
-        // Anchor: CDP scrollY value when content_origin_y was known-correct.
-        // Used to recalibrate content_origin_y against CDP absolute position
-        // and eliminate accumulated drift from content-based detection.
-        let mut cdp_scroll_anchor: Option<(i64, i64)> = None; // (cdp_scroll_y, content_origin_y)
-
-        // Send grid config so the client knows the tile layout.
-        let grid_frame = emitter.emit_grid_config(&grid);
-        if tile_tx.blocking_send(grid_frame).is_err() {
-            return;
-        }
-
-        // Optional XDamage tracking — skip frames with no changes.
-        let mut damage = capture::x11::DamageTracker::with_options(
+        if let Some(thread) = tile_loop::TileCaptureThread::new(
             &display_for_tiles,
-            None,
-            None,
-            Some(input_activity_for_tile),
-        )
-        .ok()
-        .flatten();
-        if damage.is_some() {
-            debug!("tile capture: XDamage tracking active");
+            tile_init_w,
+            tile_init_h,
+            h264_mode,
+            tile_size_from_env(),
+            tile_codec_from_env(),
+            scroll_copy_quantum_px,
+            std::time::Duration::from_millis(100),
+            std::time::Duration::from_millis(
+                env_u32_clamped("BPANE_SCROLL_ACTIVE_FRAME_INTERVAL_MS", 33, 16, 100) as u64,
+            ),
+            env_u32_clamped("BPANE_SCROLL_ACTIVE_CAPTURE_FRAMES", 8, 0, 32) as u8,
+            env_u32_clamped("BPANE_CDP_MIN_VIDEO_WIDTH", 320, 2, 4096) & !1,
+            env_u32_clamped("BPANE_CDP_MIN_VIDEO_HEIGHT", 180, 2, 4096) & !1,
+            env_f32_clamped("BPANE_CDP_MIN_VIDEO_AREA_RATIO", 0.08, 0.01, 0.95),
+            env_u32_clamped("BPANE_VIDEO_CLICK_ARM_MS", 8_000, 250, 60_000) as u64,
+            env_bool("BPANE_SCROLL_THIN_MODE", false),
+            tile_tx,
+            cmd_tx_for_tiles,
+            tiles_active_for_tiles,
+            video_tile_info_for_tiles,
+            browser_video_hint_for_tiles,
+            input_activity_for_tiles,
+            scroll_rx,
+            video_click_rx,
+            text_input_rx,
+            cache_miss_rx,
+        ) {
+            thread.run();
         }
 
-        let mut last_capture = std::time::Instant::now();
-        let mut last_resize_check = std::time::Instant::now() - std::time::Duration::from_secs(1);
-        let mut last_h264_toggle_at = std::time::Instant::now();
-
-        loop {
-            let frame_interval = select_capture_frame_interval(
-                base_frame_interval,
-                scroll_active_frame_interval,
-                scroll_active_capture_frames_remaining,
-            );
-            // Sleep until next frame interval is due (minimum 16ms for event coalescing).
-            let now = std::time::Instant::now();
-            let since_last = now.duration_since(last_capture);
-            let sleep_dur = if since_last >= frame_interval {
-                std::time::Duration::from_millis(16)
-            } else {
-                (frame_interval - since_last).max(std::time::Duration::from_millis(16))
-            };
-            std::thread::sleep(sleep_dur);
-
-            let mut force_refresh = false;
-            while let Ok((frame_seq, col, row, hash)) = cache_miss_rx.try_recv() {
-                emitter.handle_cache_miss(col, row, hash);
-                force_refresh = true;
-                trace!(
-                    frame_seq,
-                    col,
-                    row,
-                    hash,
-                    "tile cache miss reported by client"
-                );
-            }
-
-            let has_damage = match damage.as_mut() {
-                Some(dt) => dt.poll(),
-                None => true,
-            };
-            if !has_damage && !force_refresh {
-                continue;
-            }
-
-            let now = std::time::Instant::now();
-            if now.duration_since(last_capture) < frame_interval {
-                continue;
-            }
-            last_capture = now;
-
-            // Check if screen resolution changed (xrandr resize by FFmpeg pipeline).
-            // Only issue the X11 GetGeometry round-trip every ~500ms to avoid
-            // per-frame latency on the hot path. Cached values are used otherwise.
-            let (cur_w, cur_h) =
-                if now.duration_since(last_resize_check) >= std::time::Duration::from_millis(500) {
-                    last_resize_check = now;
-                    cap.refresh_screen_size()
-                } else {
-                    cap.query_screen_size()
-                };
-            if cur_w as u16 != screen_w || cur_h as u16 != screen_h {
-                screen_w = cur_w as u16;
-                screen_h = cur_h as u16;
-                grid = tiles::TileGrid::new(screen_w, screen_h, tile_size);
-                emitter = tiles::emitter::TileEmitter::with_codec(grid.cols, grid.rows, tile_codec);
-                let new_total = grid.cols as usize * grid.rows as usize;
-                prev_hashes = vec![0; new_total];
-                video_scores = vec![0; new_total];
-                non_candidate_streaks = vec![0; new_total];
-                video_hold_frames = vec![0; new_total];
-                video_latched = vec![false; new_total];
-                candidate_mask = vec![false; new_total];
-                changed_mask = vec![false; new_total];
-                text_like_mask = vec![false; new_total];
-                prev_video_bbox = None;
-                stable_bbox_frames = 0;
-                scroll_cooldown_frames = 0;
-                scroll_thin_mode_active = false;
-                scroll_residual_was_active = false;
-                scroll_quiet_frames = 0;
-                last_cdp_hint_seq = 0;
-                last_cdp_scroll_y = None;
-                cdp_scroll_anchor = None;
-                prev_frame = None;
-                content_origin_y = 0;
-                grid_offset_y = 0;
-                let had_region = region_committer.active.is_some();
-                region_committer.reset();
-                last_h264_toggle_at = std::time::Instant::now();
-                if had_region {
-                    let _ = cmd_tx_for_tile.send(capture::ffmpeg::PipelineCmd::SetRegion(None));
-                }
-                {
-                    let mut guard = match video_tile_info_for_tile.lock() {
-                        Ok(g) => g,
-                        Err(poisoned) => poisoned.into_inner(),
-                    };
-                    *guard = None;
-                }
-                debug!("tile capture: resized to {}x{}", screen_w, screen_h);
-
-                // Send new grid config
-                let grid_frame = emitter.emit_grid_config(&grid);
-                if tile_tx.blocking_send(grid_frame).is_err() {
-                    return;
-                }
-                continue;
-            }
-
-            let raw = match cap.capture_region_raw(0, 0, screen_w, screen_h) {
-                Ok(data) => data,
-                Err(e) => {
-                    warn!("tile capture: GetImage failed: {e}");
-                    continue;
-                }
-            };
-
-            // Pixels stay in native BGRA format — hashing is format-agnostic,
-            // and the per-tile BGRA→RGBA swap happens only for QOI-encoded tiles
-            // in the emitter (a small subset of total pixels).
-            let rgba = raw;
-
-            grid.advance_frame();
-            let stride = screen_w as usize * 4;
-
-            // ── Scroll displacement detection ─────────────────────
-            // Drain pending scroll events (used only for logging).
-            let mut pending_scrolls = 0i32;
-            let mut pending_scroll_dy_sum = 0i32;
-            while let Ok((_, dy)) = scroll_rx.try_recv() {
-                pending_scrolls += 1;
-                pending_scroll_dy_sum += dy as i32;
-            }
-            while let Ok((x, y, ts)) = video_click_rx.try_recv() {
-                last_left_click = Some((x, y, ts));
-            }
-            while let Ok(ts) = text_input_rx.try_recv() {
-                editable_hint.on_key_input(ts);
-            }
-
-            let prev_for_analysis = prev_frame.as_deref();
-            let cdp_hint_snapshot = {
-                let guard = match browser_video_hint_for_tile.lock() {
-                    Ok(g) => g,
-                    Err(poisoned) => poisoned.into_inner(),
-                };
-                *guard
-            };
-            let cdp_hint_region_kind = cdp_hint_snapshot.region_kind;
-            let cdp_hint_region_raw = cdp_hint_snapshot.video_region.and_then(|region| {
-                clamp_region_to_screen(region, screen_w as u32, screen_h as u32)
-            });
-            let cdp_video_region_hint_sized =
-                if matches!(cdp_hint_region_kind, cdp_video::HintRegionKind::Video) {
-                    cdp_hint_region_raw.filter(|region| {
-                        region_meets_video_minimum(
-                            region.w,
-                            region.h,
-                            screen_w as u32,
-                            screen_h as u32,
-                            min_cdp_video_width_px,
-                            min_cdp_video_height_px,
-                            min_cdp_video_area_ratio,
-                        )
-                    })
-                } else {
-                    None
-                };
-            let cdp_editable_region_hint =
-                if matches!(cdp_hint_region_kind, cdp_video::HintRegionKind::Editable) {
-                    cdp_hint_region_raw
-                        .filter(|region| region_meets_editable_minimum(region.w, region.h))
-                } else {
-                    None
-                };
-            editable_hint.update(cdp_editable_region_hint, now, EDITABLE_HINT_HOLD_MS);
-            let key_input_qoi_boost = editable_hint.key_input_qoi_boost(now, KEY_INPUT_QOI_BOOST_MS);
-            let editable_qoi_region = editable_hint.qoi_region(
-                cdp_editable_region_hint,
-                now,
-                EDITABLE_HINT_HOLD_MS,
-                KEY_INPUT_QOI_BOOST_MS,
-            );
-            let editable_qoi_tile_bounds = editable_qoi_region.map(|region| {
-                expand_tile_bounds(
-                    capture_region_tile_bounds(region, tile_size, grid.cols, grid.rows),
-                    EDITABLE_QOI_TILE_MARGIN,
-                    grid.cols,
-                    grid.rows,
-                )
-            });
-            let mut cdp_scroll_dy_px: Option<i16> = None;
-            let cdp_scale_milli = cdp_hint_snapshot.device_scale_factor_milli.max(1);
-            if let Some(scroll_y_css) = cdp_hint_snapshot.scroll_y {
-                let scroll_y = scale_css_px_to_screen_px(scroll_y_css, cdp_scale_milli);
-                if let Some(prev_scroll_y) = last_cdp_scroll_y {
-                    let dy = scroll_y.saturating_sub(prev_scroll_y);
-                    if dy != 0 {
-                        let clamped = dy.clamp(-MAX_CDP_SCROLL_DY_PX, MAX_CDP_SCROLL_DY_PX) as i16;
-                        if clamped != 0 {
-                            cdp_scroll_dy_px = Some(clamped);
-                        }
-                    }
-                }
-                last_cdp_scroll_y = Some(scroll_y);
-            } else {
-                last_cdp_scroll_y = None;
-            }
-            if cdp_scroll_dy_px.is_none()
-                && cdp_hint_snapshot.update_seq != 0
-                && cdp_hint_snapshot.update_seq != last_cdp_hint_seq
-                && cdp_hint_snapshot.scroll_delta_y != 0
-            {
-                let clamped = scale_css_px_to_screen_px(
-                    cdp_hint_snapshot.scroll_delta_y as i64,
-                    cdp_scale_milli,
-                )
-                .clamp(-MAX_CDP_SCROLL_DY_PX, MAX_CDP_SCROLL_DY_PX)
-                    as i16;
-                if clamped != 0 {
-                    cdp_scroll_dy_px = Some(clamped);
-                }
-            }
-            if cdp_hint_snapshot.update_seq != 0 {
-                last_cdp_hint_seq = cdp_hint_snapshot.update_seq;
-            }
-
-            // Content-based scroll detection: compare current frame to
-            // previous using vertical column matching. Runs on every frame
-            // (not gated on input events) so it catches all scroll sources:
-            // mouse wheel, scrollbar drag, keyboard, programmatic scrolls.
-            // Industry standard (RDP/SPICE/VNC) intercepts OS drawing
-            // commands; since we lack compositor hooks, framebuffer
-            // comparison is the correct alternative.
-            let mut strong_scroll_observed = false;
-            let mut detected_scroll_dy_px: Option<i16> = None;
-            let input_scroll_dir = (-pending_scroll_dy_sum).signum();
-            let cdp_scroll_dir = cdp_scroll_dy_px.map(|dy| (dy as i32).signum()).unwrap_or(0);
-            let hint_scroll_dir = if input_scroll_dir != 0 {
-                input_scroll_dir
-            } else {
-                cdp_scroll_dir
-            };
-            let min_scroll_dy_px = if pending_scrolls > 0 {
-                INPUT_MIN_SCROLL_DY_PX
-            } else {
-                MIN_SCROLL_DY_PX
-            };
-            let mut content_scroll: Option<(i16, f32, bool, Option<f32>)> = None;
-            let content_scroll_search_px = content_scroll_search_limit_px(cdp_scroll_dy_px);
-            if let Some(prev) = prev_for_analysis {
-                if let Some((detected_dy, confidence)) = detect_column_scroll(
-                    &rgba,
-                    prev,
-                    stride,
-                    screen_w as usize,
-                    screen_h as usize,
-                    content_scroll_search_px,
-                ) {
-                    let detected_scroll_dir = detected_dy.signum();
-                    let direction_matches = hint_scroll_dir == 0
-                        || detected_scroll_dir == 0
-                        || hint_scroll_dir == detected_scroll_dir;
-                    // Scroll hints are soft: matching direction lowers confidence threshold.
-                    let min_confidence = if hint_scroll_dir != 0 && direction_matches {
-                        INPUT_SCROLL_MIN_CONFIDENCE
-                    } else {
-                        NO_INPUT_SCROLL_MIN_CONFIDENCE
-                    };
-                    let trusted =
-                        detected_dy.abs() >= min_scroll_dy_px && confidence >= min_confidence;
-                    if trusted {
-                        content_scroll = Some((
-                            detected_dy as i16,
-                            confidence,
-                            direction_matches,
-                            Some(min_confidence),
-                        ));
-                    } else {
-                        trace!(
-                            source = "content",
-                            dy = detected_dy,
-                            confidence = format!("{:.2}", confidence),
-                            scrolls = pending_scrolls,
-                            input_dy_sum = pending_scroll_dy_sum,
-                            input_dir = input_scroll_dir,
-                            cdp_dir = cdp_scroll_dir,
-                            dir_match = direction_matches,
-                            min_scroll_dy_px,
-                            min_confidence = format!("{:.2}", min_confidence),
-                            "ignored tiny scroll displacement"
-                        );
-                    }
-                }
-            }
-            let mut trusted_scroll: Option<(i16, f32, &'static str, bool, Option<f32>)> = None;
-            let mut detected_scroll_frame: Option<(
-                i16,
-                f32,
-                &'static str,
-                bool,
-                Option<f32>,
-                i32,
-                u16,
-                u16,
-                u16,
-            )> = None;
-
-            if let Some(detected_dy) = cdp_scroll_dy_px {
-                let cdp_abs = (detected_dy as i32).abs();
-                if cdp_abs >= min_scroll_dy_px {
-                    if pending_scrolls > 0 {
-                        if let Some((
-                            selected_dy,
-                            selected_confidence,
-                            source,
-                            direction_matches,
-                            min_confidence,
-                        )) = select_wheel_trusted_scroll(
-                            detected_dy,
-                            input_scroll_dir,
-                            content_scroll,
-                        ) {
-                            if source == "content" {
-                                let dy_gap = ((selected_dy as i32) - (detected_dy as i32)).abs();
-                                if dy_gap >= CDP_CONTENT_DY_DIVERGENCE_LOG_PX {
-                                    trace!(
-                                        cdp_dy = detected_dy,
-                                        content_dy = selected_dy,
-                                        dy_gap,
-                                        confidence = format!("{:.2}", selected_confidence),
-                                        "wheel scroll dy diverged; preferring content (pixel-aligned)"
-                                    );
-                                }
-                            }
-                            trusted_scroll = Some((
-                                selected_dy,
-                                selected_confidence,
-                                source,
-                                direction_matches,
-                                min_confidence,
-                            ));
-                        } else {
-                            let detected_scroll_dir = (detected_dy as i32).signum();
-                            let direction_matches = input_scroll_dir == 0
-                                || detected_scroll_dir == 0
-                                || input_scroll_dir == detected_scroll_dir;
-                            trace!(
-                                source = "cdp",
-                                dy = detected_dy,
-                                scrolls = pending_scrolls,
-                                input_dy_sum = pending_scroll_dy_sum,
-                                input_dir = input_scroll_dir,
-                                dir_match = direction_matches,
-                                min_scroll_dy_px,
-                                "ignored cdp scroll hint with mismatched direction"
-                            );
-                        }
-                    } else if let Some((
-                        content_dy,
-                        content_confidence,
-                        content_direction_matches,
-                        content_min_confidence,
-                    )) = content_scroll
-                    {
-                        let cdp_dir = (detected_dy as i32).signum();
-                        let content_dir = (content_dy as i32).signum();
-                        if cdp_dir == 0 || content_dir == 0 || cdp_dir == content_dir {
-                            let dy_gap = ((content_dy as i32) - (detected_dy as i32)).abs();
-                            if dy_gap >= CDP_CONTENT_DY_DIVERGENCE_LOG_PX {
-                                trace!(
-                                    cdp_dy = detected_dy,
-                                    content_dy,
-                                    dy_gap,
-                                    confidence = format!("{:.2}", content_confidence),
-                                    "cdp/content scroll dy diverged; preferring content (pixel-aligned)"
-                                );
-                            }
-                            // Prefer content-based dy: it is derived from the
-                            // actual captured frames and therefore pixel-aligned
-                            // with the residual comparison.  CDP dy comes from a
-                            // different timing domain and may not match the
-                            // framebuffer state at capture time.
-                            trusted_scroll = Some((
-                                content_dy,
-                                content_confidence,
-                                "content",
-                                content_direction_matches,
-                                content_min_confidence,
-                            ));
-                        } else {
-                            trace!(
-                                cdp_dy = detected_dy,
-                                content_dy,
-                                confidence = format!("{:.2}", content_confidence),
-                                "cdp/content direction mismatch; preferring content"
-                            );
-                            trusted_scroll = Some((
-                                content_dy,
-                                content_confidence,
-                                "content",
-                                content_direction_matches,
-                                content_min_confidence,
-                            ));
-                        }
-                    } else {
-                        // CDP reports a scroll but content-based detection cannot
-                        // confirm it in the captured pixels.  This typically means
-                        // the frame was captured mid-render — the browser's scrollY
-                        // has changed but the framebuffer doesn't yet reflect the
-                        // full shift.  Entering scroll mode here would use a delta
-                        // that doesn't match the actual pixels, causing the residual
-                        // comparison to fail and triggering a costly full repaint.
-                        //
-                        // Instead, skip scroll mode and emit only the XDamage dirty
-                        // tiles as a normal frame.  The next frame capture will
-                        // likely see the completed render and content detection will
-                        // confirm the scroll then.
-                        trace!(
-                            source = "cdp-passive-skipped",
-                            dy = detected_dy,
-                            scrolls = pending_scrolls,
-                            cdp_dir = cdp_scroll_dir,
-                            "cdp scroll unconfirmed by content detection; skipping scroll mode"
-                        );
-                    }
-                } else if pending_scrolls > 0 {
-                    trace!(
-                        source = "cdp",
-                        dy = detected_dy,
-                        scrolls = pending_scrolls,
-                        input_dy_sum = pending_scroll_dy_sum,
-                        input_dir = input_scroll_dir,
-                        min_scroll_dy_px,
-                        "ignored tiny cdp scroll hint"
-                    );
-                }
-            }
-
-            if trusted_scroll.is_none() {
-                if let Some((dy, confidence, direction_matches, min_confidence)) = content_scroll {
-                    trusted_scroll =
-                        Some((dy, confidence, "content", direction_matches, min_confidence));
-                }
-            }
-
-            if let Some((detected_dy, confidence, source, direction_matches, min_confidence)) =
-                trusted_scroll
-            {
-                content_origin_y += detected_dy as i64;
-                // Re-anchor after each trusted scroll so drift correction stays fresh.
-                if let Some(cdp_scroll_y_css) = cdp_hint_snapshot.scroll_y {
-                    let cdp_scroll_y = scale_css_px_to_screen_px(cdp_scroll_y_css, cdp_scale_milli);
-                    cdp_scroll_anchor = Some((cdp_scroll_y, content_origin_y));
-                }
-                grid_offset_y = 0;
-
-                // Keep tiles in fixed screen-space. ScrollCopy is only used
-                // for whole-tile moves, so row-shift is derived directly from
-                // the observed scroll delta rather than a moving tile grid.
-                let row_shift = if detected_dy as i32 % tile_size as i32 == 0 {
-                    detected_dy as i32 / tile_size as i32
-                } else {
-                    0
-                };
-
-                // Determine scroll region from CDP viewport hint.
-                // If available, limit canvas shift to viewport content area
-                // so browser toolbar and scrollbar don't jump.
-                let (scroll_region_top, scroll_region_bottom, scroll_region_right) = {
-                    if let Some(vp) = cdp_hint_snapshot.viewport {
-                        (
-                            (vp.y as u16).min(screen_h),
-                            ((vp.y + vp.h) as u16).min(screen_h),
-                            ((vp.x + vp.w) as u16).min(screen_w),
-                        )
-                    } else {
-                        (0, screen_h, screen_w)
-                    }
-                };
-
-                detected_scroll_frame = Some((
-                    detected_dy,
-                    confidence,
-                    source,
-                    direction_matches,
-                    min_confidence,
-                    row_shift,
-                    scroll_region_top,
-                    scroll_region_bottom,
-                    scroll_region_right,
-                ));
-                strong_scroll_observed = true;
-                detected_scroll_dy_px = Some(detected_dy);
-            } else {
-                // Recalibrate content_origin_y against CDP absolute scrollY
-                // to eliminate accumulated drift from content-based detection.
-                // This ensures scroll-back to a previous position produces
-                // identical grid_offset_y → identical tile hashes → cache hits.
-                //
-                // IMPORTANT: only apply when no scroll is detected this frame.
-                // If both CDP correction and scroll detection fire in the same
-                // frame, the scroll delta is double-counted — CDP correction
-                // adds the delta, then scroll detection adds it again — causing
-                // grid_offset_y to be 2x off and producing ghost/double text.
-                if let Some(cdp_scroll_y_css) = cdp_hint_snapshot.scroll_y {
-                    let cdp_scroll_y = scale_css_px_to_screen_px(cdp_scroll_y_css, cdp_scale_milli);
-                    if let Some((anchor_scroll_y, anchor_origin)) = cdp_scroll_anchor {
-                        let expected_origin = anchor_origin + (cdp_scroll_y - anchor_scroll_y);
-                        let drift = content_origin_y - expected_origin;
-                        if drift != 0 {
-                            tracing::trace!(
-                                drift,
-                                content_origin_y,
-                                expected_origin,
-                                cdp_scroll_y,
-                                "correcting content_origin_y drift from CDP"
-                            );
-                            content_origin_y = expected_origin;
-                            grid_offset_y = 0;
-                        }
-                    } else {
-                        // Establish anchor on first CDP reading
-                        cdp_scroll_anchor = Some((cdp_scroll_y, content_origin_y));
-                    }
-                }
-            }
-            let cdp_scroll_observed = cdp_scroll_dy_px.is_some();
-            scroll_active_capture_frames_remaining = next_scroll_active_capture_frames(
-                scroll_active_capture_frames_remaining,
-                scroll_active_capture_frames,
-                pending_scrolls,
-                strong_scroll_observed,
-                cdp_scroll_observed,
-            );
-            if pending_scrolls > 0 || strong_scroll_observed || cdp_scroll_observed {
-                // Scroll-like motion always wins over video classification.
-                // Keep video suppressed for a short quiet period.
-                scroll_cooldown_frames = SCROLL_SUPPRESS_VIDEO_FRAMES;
-                stable_bbox_frames = 0;
-                prev_video_bbox = None;
-                scroll_quiet_frames = 0;
-            } else if scroll_cooldown_frames > 0 {
-                scroll_cooldown_frames -= 1;
-                scroll_quiet_frames = scroll_quiet_frames.saturating_add(1);
-            } else {
-                scroll_quiet_frames = scroll_quiet_frames.saturating_add(1);
-            }
-            let cdp_video_region_hint_candidate = if scroll_cooldown_frames == 0 {
-                cdp_video_region_hint_sized
-            } else {
-                None
-            };
-            let cdp_click_armed = click_armed.update(
-                cdp_video_region_hint_candidate,
-                last_left_click,
-                now,
-                video_click_arm_ms,
-                CLICK_LATCH_RESET_FRAMES,
-            );
-            let cdp_video_region_hint = if cdp_click_armed {
-                cdp_video_region_hint_candidate
-            } else {
-                None
-            };
-            let cdp_hint_tile_bounds = cdp_video_region_hint
-                .map(|region| capture_region_tile_bounds(region, tile_size, grid.cols, grid.rows));
-
-            // ── Per-tile change detection (two-pass) ─────────────
-            // Pass 1: hash each tile, extract motion features for changed
-            // tiles, identify video candidates, compute candidate bbox.
-            let cols = grid.cols as usize;
-            candidate_mask.fill(false);
-            changed_mask.fill(false);
-            text_like_mask.fill(false);
-            let mut bbox_min_col: u16 = u16::MAX;
-            let mut bbox_max_col: u16 = 0;
-            let mut bbox_min_row: u16 = u16::MAX;
-            let mut bbox_max_row: u16 = 0;
-            let mut candidate_count = 0u32;
-            let mut strong_candidate_count = 0u32;
-
-            for row in 0..grid.rows {
-                for col in 0..grid.cols {
-                    let idx = row as usize * cols + col as usize;
-                    let tx = col as usize * tile_size as usize;
-                    let ty = row as usize * tile_size as usize;
-                    let tw = (tile_size as usize).min(screen_w as usize - tx);
-                    let th = (tile_size as usize).min(screen_h as usize - ty);
-                    let hash = hash_tile_region(&rgba, stride, tx, ty, tw, th);
-                    let changed = prev_hashes[idx] != 0 && hash != prev_hashes[idx];
-                    changed_mask[idx] = changed;
-
-                    if changed
-                        && video_classification_enabled
-                        && cdp_video_region_hint.is_some()
-                        && scroll_cooldown_frames == 0
-                    {
-                        let features = compute_tile_motion_features(
-                            &rgba,
-                            prev_for_analysis,
-                            stride,
-                            tx,
-                            ty,
-                            tw,
-                            th,
-                        );
-                        // Text/UI is typically edge-heavy and lower entropy.
-                        // Video/canvas is usually high motion + richer entropy.
-                        text_like_mask[idx] =
-                            features.edge_density > 0.22 && features.entropy_hint < 0.45;
-                        let photo_like = is_photo_like_tile(&rgba, stride, tx, ty, tw, th, 16);
-                        let video_like = features.change_ratio > 0.23
-                            && features.motion_magnitude > 0.045
-                            && features.entropy_hint > 0.16
-                            && features.edge_density < 0.74
-                            && photo_like;
-                        let strong_video_like = features.change_ratio > 0.30
-                            && features.motion_magnitude > 0.07
-                            && features.entropy_hint > 0.20
-                            && features.edge_density < 0.70
-                            && photo_like;
-
-                        if video_like {
-                            candidate_mask[idx] = true;
-                            candidate_count += 1;
-                            if strong_video_like {
-                                strong_candidate_count += 1;
-                            }
-                            bbox_min_col = bbox_min_col.min(col);
-                            bbox_max_col = bbox_max_col.max(col);
-                            bbox_min_row = bbox_min_row.min(row);
-                            bbox_max_row = bbox_max_row.max(row);
-                        }
-                    }
-                    prev_hashes[idx] = hash;
-                }
-            }
-
-            // Bounding box stability check:
-            // Video regions stay roughly stable in position/shape, while
-            // scrolling tends to translate the bbox frame to frame.
-            let current_bbox = if candidate_count > 0 {
-                Some((bbox_min_col, bbox_min_row, bbox_max_col, bbox_max_row))
-            } else {
-                None
-            };
-
-            let bbox_stable = match (current_bbox, prev_video_bbox) {
-                (Some(cur), Some(prev)) => {
-                    let iou = bbox_iou(cur, prev);
-                    let center_shift = bbox_center_shift(cur, prev);
-                    iou > 0.55 || center_shift <= 1.25
-                }
-                _ => false,
-            };
-            if bbox_stable {
-                stable_bbox_frames = stable_bbox_frames.saturating_add(1);
-            } else {
-                stable_bbox_frames = 0;
-            }
-            prev_video_bbox = current_bbox;
-            let bbox_tile_area = current_bbox
-                .map(|(min_c, min_r, max_c, max_r)| {
-                    ((max_c - min_c + 1) as u32).saturating_mul((max_r - min_r + 1) as u32)
-                })
-                .unwrap_or(0);
-            let bbox_density = if bbox_tile_area > 0 {
-                candidate_count as f32 / bbox_tile_area as f32
-            } else {
-                0.0
-            };
-            let (bbox_w_tiles, bbox_h_tiles) = current_bbox
-                .map(|(min_c, min_r, max_c, max_r)| (max_c - min_c + 1, max_r - min_r + 1))
-                .unwrap_or((0, 0));
-            let total_tile_area = (grid.cols as u32).saturating_mul(grid.rows as u32).max(1);
-            let bbox_area_ratio = bbox_tile_area as f32 / total_tile_area as f32;
-            let large_stable_region = bbox_w_tiles >= MIN_VIDEO_BBOX_WIDTH_TILES
-                && bbox_h_tiles >= MIN_VIDEO_BBOX_HEIGHT_TILES
-                && bbox_area_ratio >= MIN_VIDEO_BBOX_AREA_RATIO;
-            let region_stable = video_classification_enabled
-                && cdp_video_region_hint.is_some()
-                && scroll_cooldown_frames == 0
-                && candidate_count >= REGION_MIN_CANDIDATES
-                && large_stable_region
-                && ((stable_bbox_frames >= 4 && bbox_density > 0.34)
-                    || candidate_count >= REGION_DENSE_CANDIDATES
-                    || strong_candidate_count >= 8);
-
-            // Pass 2: update per-tile motion scores + hysteresis and classify.
-            let mut latched_video_tiles: Vec<tiles::TileCoord> = Vec::new();
-            let mut cdp_motion_tiles: u32 = 0;
-            for row in 0..grid.rows {
-                for col in 0..grid.cols {
-                    let idx = row as usize * cols + col as usize;
-                    let in_cdp_video_hint = cdp_hint_tile_bounds
-                        .map(|(min_col, min_row, max_col, max_row)| {
-                            col >= min_col && col <= max_col && row >= min_row && row <= max_row
-                        })
-                        .unwrap_or(false);
-                    let cdp_motion_candidate =
-                        in_cdp_video_hint && changed_mask[idx] && !text_like_mask[idx];
-                    if cdp_motion_candidate {
-                        cdp_motion_tiles = cdp_motion_tiles.saturating_add(1);
-                    }
-                    if region_stable && candidate_mask[idx] {
-                        video_scores[idx] = (video_scores[idx] + 4).min(VIDEO_MAX_SCORE);
-                        non_candidate_streaks[idx] = 0;
-                        if video_latched[idx] {
-                            video_hold_frames[idx] = VIDEO_MIN_HOLD_FRAMES;
-                        }
-                    } else if cdp_motion_candidate {
-                        video_scores[idx] = (video_scores[idx] + 2).min(VIDEO_MAX_SCORE);
-                        non_candidate_streaks[idx] = 0;
-                        if video_latched[idx] {
-                            video_hold_frames[idx] = VIDEO_MIN_HOLD_FRAMES;
-                        }
-                    } else if changed_mask[idx] {
-                        video_scores[idx] = (video_scores[idx] - 1).max(-VIDEO_MAX_SCORE);
-                        non_candidate_streaks[idx] = non_candidate_streaks[idx].saturating_add(1);
-                    } else {
-                        video_scores[idx] = (video_scores[idx] - 1).max(-VIDEO_MAX_SCORE);
-                        non_candidate_streaks[idx] = non_candidate_streaks[idx].saturating_add(1);
-                    }
-
-                    if !(region_stable && candidate_mask[idx] || cdp_motion_candidate)
-                        && video_hold_frames[idx] > 0
-                    {
-                        video_hold_frames[idx] -= 1;
-                    }
-
-                    if !video_classification_enabled
-                        || cdp_video_region_hint.is_none()
-                        || scroll_cooldown_frames > 0
-                    {
-                        video_latched[idx] = false;
-                        video_hold_frames[idx] = 0;
-                        video_scores[idx] = video_scores[idx].min(0);
-                    } else if video_latched[idx] {
-                        if video_hold_frames[idx] == 0
-                            && (video_scores[idx] <= VIDEO_EXIT_SCORE
-                                || non_candidate_streaks[idx] >= VIDEO_DECAY_STREAK)
-                        {
-                            video_latched[idx] = false;
-                            video_hold_frames[idx] = 0;
-                        }
-                    } else if (region_stable && video_scores[idx] >= VIDEO_ENTER_SCORE)
-                        || (cdp_motion_candidate && video_scores[idx] >= (VIDEO_ENTER_SCORE - 1))
-                    {
-                        video_latched[idx] = true;
-                        video_hold_frames[idx] = VIDEO_MIN_HOLD_FRAMES;
-                    }
-
-                    if let Some(tile) = grid.get_mut(tiles::TileCoord::new(col, row)) {
-                        tile.dirty = true;
-                        tile.classification = if video_latched[idx] {
-                            tiles::TileClass::VideoMotion
-                        } else if changed_mask[idx] && text_like_mask[idx] {
-                            tiles::TileClass::TextScroll
-                        } else {
-                            tiles::TileClass::Static
-                        };
-                    }
-                    if video_latched[idx] {
-                        latched_video_tiles.push(tiles::TileCoord::new(col, row));
-                    }
-                }
-            }
-            // Emit tiles for all positions, including extra bottom row when
-            // grid offset is active (partial tile at bottom edge).
-            let emit_rows = if grid_offset_y > 0 {
-                grid.rows + 1
-            } else {
-                grid.rows
-            };
-            let full_emit_coords: Vec<tiles::TileCoord> = (0..emit_rows)
-                .flat_map(|r| (0..grid.cols).map(move |c| tiles::TileCoord::new(c, r)))
-                .collect();
-            // Narrow dirty set to tiles overlapping XDamage bounding box.
-            // Tiles outside the damage region are guaranteed unchanged by the
-            // X server, so we skip their extraction + hashing entirely.
-            // During scroll or force-refresh we fall back to all tiles.
-            let mut all_dirty: Vec<tiles::TileCoord> = if !force_refresh {
-                if let Some(ref dt) = damage {
-                    if let Some((dx, dy, dw, dh)) = dt.damage_bounding_box() {
-                        let ts = tile_size as u16;
-                        let dx2 = dx.saturating_add(dw);
-                        let dy2 = dy.saturating_add(dh);
-                        full_emit_coords
-                            .iter()
-                            .copied()
-                            .filter(|coord| {
-                                let tx = coord.col * ts;
-                                let ty = coord.row * ts;
-                                let tx2 = tx.saturating_add(ts);
-                                let ty2 = ty.saturating_add(ts);
-                                // AABB overlap test
-                                tx < dx2 && tx2 > dx && ty < dy2 && ty2 > dy
-                            })
-                            .collect()
-                    } else {
-                        full_emit_coords.clone()
-                    }
-                } else {
-                    full_emit_coords.clone()
-                }
-            } else {
-                full_emit_coords.clone()
-            };
-            let mut scroll_residual_ratio: Option<f32> = None;
-            let mut scroll_residual_fallback_full = false;
-            let mut scroll_residual_tiles_frame: Option<usize> = None;
-            let mut scroll_potential_tiles_frame: Option<usize> = None;
-            let mut scroll_saved_tiles_frame: Option<usize> = None;
-            let mut scroll_saved_ratio_frame: Option<f32> = None;
-            let mut scroll_emit_ratio_frame: Option<f32> = None;
-            let mut scroll_thin_mode_frame = false;
-            let mut scroll_thin_repair_frame = false;
-            if let (Some(scroll_dy), Some(prev)) = (detected_scroll_dy_px, prev_for_analysis) {
-                let scroll_row_shift = detected_scroll_frame
-                    .map(|(_, _, _, _, _, row_shift, _, _, _)| row_shift)
-                    .unwrap_or(0);
-                // Get scroll region bounds for partitioning tiles into
-                // content (scrollable) and chrome (static).  Use freshly
-                // detected values when available, fall back to cached.
-                let srt_for_split = detected_scroll_frame
-                    .map(|(_, _, _, _, _, _, srt, _, _)| srt)
-                    .unwrap_or(last_scroll_region_top);
-                let srb_for_split = detected_scroll_frame
-                    .map(|(_, _, _, _, _, _, _, srb, _)| srb)
-                    .unwrap_or(screen_h);
-                let srr_for_split = detected_scroll_frame
-                    .map(|(_, _, _, _, _, _, _, _, srr)| srr)
-                    .unwrap_or(last_scroll_region_right);
-                let ts = tile_size as u16;
-
-                // Partition: content tiles are below chrome header and left
-                // of scrollbar; everything else is chrome / static.
-                // Chrome exists regardless of sub-tile offset; partition
-                // whenever we have a detected scroll region.
-                let have_split = has_scroll_region_split(
-                    srt_for_split,
-                    srb_for_split,
-                    srr_for_split,
-                    screen_h,
-                    screen_w,
-                );
-                let (content_emit_coords, chrome_emit_coords): (
-                    Vec<tiles::TileCoord>,
-                    Vec<tiles::TileCoord>,
-                ) = if have_split {
-                    full_emit_coords.iter().partition(|coord| {
-                        // Only tiles fully inside the scrollable viewport are
-                        // eligible for ScrollCopy reuse. Boundary tiles that
-                        // overlap the header, bottom seam, or scrollbar remain
-                        // raw/static so the host never assumes partially moved
-                        // tiles are already correct on the client.
-                        is_content_tile_in_scroll_region(
-                            **coord,
-                            ts,
-                            srt_for_split,
-                            srb_for_split,
-                            srr_for_split,
-                        )
-                    })
-                } else {
-                    (full_emit_coords.clone(), Vec::new())
-                };
-
-                let residual_coords = content_emit_coords.clone();
-                let residual = build_scroll_residual_emit_coords(
-                    &rgba,
-                    prev,
-                    stride,
-                    &grid,
-                    grid_offset_y,
-                    scroll_dy,
-                    &residual_coords,
-                );
-
-                let potential_tiles = residual_coords.len();
-                let residual_tiles = residual.len();
-                let residual_ratio = if potential_tiles == 0 {
-                    1.0
-                } else {
-                    residual_tiles as f32 / potential_tiles as f32
-                };
-                scroll_residual_batches_total = scroll_residual_batches_total.saturating_add(1);
-                scroll_potential_tiles_total =
-                    scroll_potential_tiles_total.saturating_add(potential_tiles as u64);
-                scroll_residual_tiles_total =
-                    scroll_residual_tiles_total.saturating_add(residual_tiles as u64);
-                scroll_residual_tiles_frame = Some(residual_tiles);
-                scroll_potential_tiles_frame = Some(potential_tiles);
-                scroll_residual_ratio = Some(residual_ratio);
-
-                // Fallback decision: only triggers if the residual ratio
-                // for INTERIOR content tiles (excluding newly exposed edges)
-                // exceeds the threshold, indicating scroll detection inaccuracy.
-                let exposed_tiles = build_scroll_exposed_strip_emit_coords(
-                    &grid,
-                    grid_offset_y,
-                    scroll_dy,
-                    &residual_coords,
-                );
-                let interior_residual = residual
-                    .iter()
-                    .filter(|c| !exposed_tiles.contains(c))
-                    .count();
-                let interior_total = potential_tiles.saturating_sub(exposed_tiles.len());
-                let interior_ratio = if interior_total == 0 {
-                    0.0
-                } else {
-                    interior_residual as f32 / interior_total as f32
-                };
-                let quantized_scroll_copy =
-                    can_emit_scroll_copy(scroll_dy, scroll_copy_quantum_px_for_tile, tile_size);
-                let saved_tiles = potential_tiles.saturating_sub(residual_tiles);
-                let defer_scroll_repair = should_defer_scroll_repair(
-                    quantized_scroll_copy,
-                    interior_ratio,
-                    saved_tiles,
-                    potential_tiles,
-                    scroll_row_shift,
-                );
-                if !quantized_scroll_copy {
-                    trace!(
-                        dy = scroll_dy,
-                        scroll_copy_quantum_px = scroll_copy_quantum_px_for_tile,
-                        "scroll copy suppressed for non-quantized delta"
-                    );
-                    scroll_residual_fallback_full = true;
-                    scroll_residual_fallback_full_total =
-                        scroll_residual_fallback_full_total.saturating_add(1);
-                    scroll_saved_tiles_frame = Some(0);
-                    scroll_saved_ratio_frame = Some(0.0);
-                    scroll_emit_ratio_frame = Some(1.0);
-                    scroll_thin_mode_active = false;
-                    scroll_residual_was_active = false;
-                    // Restore full dirty set — XDamage narrowing may have
-                    // excluded tiles that need a full repaint after scroll.
-                    all_dirty = full_emit_coords.clone();
-                } else if interior_ratio > SCROLL_RESIDUAL_FULL_REPAINT_RATIO
-                    && !defer_scroll_repair
-                {
-                    trace!(
-                        dy = scroll_dy,
-                        row_shift = scroll_row_shift,
-                        interior_ratio = format!("{:.2}", interior_ratio),
-                        saved_tiles,
-                        potential_tiles,
-                        "scroll copy suppressed by residual full repaint threshold"
-                    );
-                    scroll_residual_fallback_full = true;
-                    scroll_residual_fallback_full_total =
-                        scroll_residual_fallback_full_total.saturating_add(1);
-                    scroll_saved_tiles_frame = Some(0);
-                    scroll_saved_ratio_frame = Some(0.0);
-                    scroll_emit_ratio_frame = Some(1.0);
-                    scroll_thin_mode_active = false;
-                    scroll_residual_was_active = false;
-                    // Restore full dirty set — XDamage narrowing may have
-                    // excluded tiles that need a full repaint after scroll.
-                    all_dirty = full_emit_coords.clone();
-                } else {
-                    // Dirty set = residual content + chrome.
-                    all_dirty = residual;
-                    all_dirty.extend(chrome_emit_coords.iter().copied());
-
-                    // Force content tiles overlapping the client-side exposed
-                    // strip into the dirty set.  The residual analysis compares
-                    // pixels (white == white on a uniform page) so these tiles
-                    // pass the check, but ScrollCopy cleared their canvas region.
-                    // Without this, they'd be skipped → black band.
-                    {
-                        let (exp_start, exp_end) = if scroll_dy > 0 {
-                            let start =
-                                (srb_for_split as i32 - scroll_dy as i32).max(srt_for_split as i32);
-                            (start, srb_for_split as i32)
-                        } else {
-                            let end = (srt_for_split as i32 + (-scroll_dy) as i32)
-                                .min(srb_for_split as i32);
-                            (srt_for_split as i32, end)
-                        };
-                        for &coord in &content_emit_coords {
-                            if all_dirty.contains(&coord) {
-                                continue;
-                            }
-                            let rect = offset_tile_rect_for_emit(coord, &grid, grid_offset_y);
-                            if rect.w == 0 || rect.h == 0 {
-                                continue;
-                            }
-                            let tile_top = rect.y as i32;
-                            let tile_bot = tile_top + rect.h as i32;
-                            if tile_bot > exp_start && tile_top < exp_end {
-                                all_dirty.push(coord);
-                            }
-                        }
-                    }
-                    scroll_saved_tiles_total =
-                        scroll_saved_tiles_total.saturating_add(saved_tiles as u64);
-                    scroll_saved_tiles_frame = Some(saved_tiles);
-                    if potential_tiles > 0 {
-                        scroll_saved_ratio_frame =
-                            Some(saved_tiles as f32 / potential_tiles as f32);
-                        scroll_emit_ratio_frame =
-                            Some(residual_tiles as f32 / potential_tiles as f32);
-                    } else {
-                        scroll_saved_ratio_frame = Some(0.0);
-                        scroll_emit_ratio_frame = Some(1.0);
-                    }
-                    scroll_residual_was_active = saved_tiles > 0;
-                    if defer_scroll_repair {
-                        trace!(
-                            dy = scroll_dy,
-                            row_shift = scroll_row_shift,
-                            interior_ratio = format!("{:.2}", interior_ratio),
-                            saved_tiles,
-                            potential_tiles,
-                            "scroll copy accepted with deferred repair"
-                        );
-                    }
-
-                    // If sub-tile residual explodes, prioritize exposed strip
-                    // during active scrolling and defer one full repair frame.
-                    let sub_tile_scroll = (scroll_dy.unsigned_abs() as u16) < tile_size;
-                    let residual_tiles_min_for_thin = (grid.cols as usize * 2).max(12);
-                    let residual_large_for_sub_tile = residual_ratio
-                        >= SCROLL_THIN_MODE_RESIDUAL_RATIO
-                        || residual_tiles >= residual_tiles_min_for_thin;
-                    if sub_tile_scroll && residual_large_for_sub_tile {
-                        let strip_dirty = build_scroll_exposed_strip_emit_coords(
-                            &grid,
-                            grid_offset_y,
-                            scroll_dy,
-                            &residual_coords,
-                        );
-                        if scroll_thin_mode_enabled && !strip_dirty.is_empty() {
-                            // Keep chrome tiles in dirty set for static emit.
-                            all_dirty = strip_dirty;
-                            all_dirty.extend(chrome_emit_coords.iter().copied());
-                            scroll_thin_mode_active = true;
-                            scroll_thin_mode_frame = true;
-                            scroll_thin_batches_total = scroll_thin_batches_total.saturating_add(1);
-                        } else {
-                            scroll_thin_mode_active = false;
-                        }
-                    } else {
-                        scroll_thin_mode_active = false;
-                    }
-                }
-            } else if (scroll_thin_mode_active || scroll_residual_was_active)
-                && scroll_quiet_frames >= SCROLL_THIN_REPAIR_QUIET_FRAMES
-            {
-                // Reconcile: after scroll quiets, force a full tile emit to
-                // correct any accumulated errors from ScrollCopy + residual skipping.
-                // This ensures tiles skipped by residual analysis (whose last_hashes
-                // are stale) get properly re-evaluated and updated.
-                all_dirty = full_emit_coords.clone();
-                scroll_thin_mode_active = false;
-                scroll_residual_was_active = false;
-                scroll_thin_repair_frame = true;
-                scroll_thin_repairs_total = scroll_thin_repairs_total.saturating_add(1);
-            }
-
-            let emit_scroll_copy =
-                should_emit_scroll_copy(scroll_residual_fallback_full, scroll_saved_tiles_frame);
-            if let Some((
-                detected_dy,
-                confidence,
-                source,
-                direction_matches,
-                min_confidence,
-                row_shift,
-                scroll_region_top,
-                scroll_region_bottom,
-                scroll_region_right,
-            )) = detected_scroll_frame
-            {
-                if emit_scroll_copy {
-                    // Only shift hashes when ScrollCopy is actually sent —
-                    // keeps last_hashes consistent with the client canvas.
-                    if row_shift != 0 {
-                        emitter.shift_hashes(row_shift, grid.rows);
-                    }
-                    // Always zero exposed strip when sending ScrollCopy,
-                    // even for sub-tile scrolls (row_shift == 0).  The client
-                    // keeps that strip stale until repair tiles arrive; if we
-                    // don't zero the corresponding hashes, L1 skip can keep
-                    // the stale strip visible indefinitely.
-                    emitter.zero_exposed_strip(
-                        detected_dy,
-                        scroll_region_top,
-                        scroll_region_bottom,
-                        tile_size,
-                        grid_offset_y,
-                    );
-                    let scroll_frame = bpane_protocol::TileMessage::ScrollCopy {
-                        dx: 0,
-                        dy: detected_dy,
-                        region_top: scroll_region_top,
-                        region_bottom: scroll_region_bottom,
-                        region_right: scroll_region_right,
-                    }
-                    .to_frame();
-                    if tile_tx.blocking_send(scroll_frame).is_err() {
-                        return;
-                    }
-                }
-
-                let offset_frame = bpane_protocol::TileMessage::GridOffset {
-                    offset_x: 0,
-                    offset_y: grid_offset_y as i16,
-                }
-                .to_frame();
-                if tile_tx.blocking_send(offset_frame).is_err() {
-                    return;
-                }
-
-                tracing::debug!(
-                    source,
-                    dy = detected_dy,
-                    confidence = format!("{:.2}", confidence),
-                    offset_y = grid_offset_y,
-                    row_shift = row_shift,
-                    scroll_copy = emit_scroll_copy,
-                    scrolls = pending_scrolls,
-                    input_dy_sum = pending_scroll_dy_sum,
-                    input_dir = input_scroll_dir,
-                    cdp_dir = cdp_scroll_dir,
-                    dir_match = direction_matches,
-                    min_scroll_dy_px,
-                    min_confidence = min_confidence.map(|c| format!("{:.2}", c)),
-                    "scroll detected"
-                );
-            }
-
-            // Export cumulative host-side residual telemetry to client for
-            // direct Scroll Health reporting in test dashboards.
-            let to_u32_sat = |v: u64| -> u32 { v.min(u32::MAX as u64) as u32 };
-            let scroll_stats_frame = bpane_protocol::TileMessage::ScrollStats {
-                scroll_batches_total: to_u32_sat(scroll_residual_batches_total),
-                scroll_full_fallbacks_total: to_u32_sat(scroll_residual_fallback_full_total),
-                scroll_potential_tiles_total: to_u32_sat(scroll_potential_tiles_total),
-                scroll_saved_tiles_total: to_u32_sat(scroll_saved_tiles_total),
-            }
-            .to_frame();
-            if tile_tx.blocking_send(scroll_stats_frame).is_err() {
-                return;
-            }
-
-            // Split dirty tiles into static (browser chrome) and content (scrolling
-            // viewport). Static tiles are emitted at raw framebuffer positions
-            // with a separate hash table that is never shifted, so browser
-            // header/scrollbar tiles achieve L1 cache hits across scroll frames.
-            if let Some((_, _, _, _, _, _, srt, srb, srr)) = detected_scroll_frame {
-                last_scroll_region_top = srt;
-                last_scroll_region_bottom = srb;
-                last_scroll_region_right = srr;
-            }
-            if key_input_qoi_boost {
-                if let Some(bounds) = editable_qoi_tile_bounds {
-                    extend_dirty_with_tile_bounds(&mut all_dirty, bounds);
-                }
-            }
-            let use_static_split = has_scroll_region_split(
-                last_scroll_region_top,
-                last_scroll_region_bottom,
-                last_scroll_region_right,
-                screen_h,
-                screen_w,
-            );
-            let (content_dirty, static_dirty): (Vec<tiles::TileCoord>, Vec<tiles::TileCoord>) =
-                if use_static_split {
-                    let ts = tile_size as u16;
-                    let srt = last_scroll_region_top;
-                    let srb = last_scroll_region_bottom;
-                    let srr = last_scroll_region_right;
-                    all_dirty.iter().partition(|coord| {
-                        // Must stay in sync with the residual-analysis
-                        // partitioning above.
-                        is_content_tile_in_scroll_region(**coord, ts, srt, srb, srr)
-                    })
-                } else {
-                    (all_dirty, Vec::new())
-                };
-
-            let result = emitter.emit_frame(
-                &rgba,
-                stride,
-                &content_dirty,
-                &grid,
-                grid_offset_y,
-                editable_qoi_tile_bounds,
-            );
-
-            // Emit static (chrome) tiles separately with offset_y=0.
-            let static_frames = if !static_dirty.is_empty() {
-                let draw_mode_off = bpane_protocol::TileMessage::TileDrawMode {
-                    apply_offset: false,
-                }
-                .to_frame();
-                let ts = tile_size as u16;
-                let boundary_col = if use_static_split
-                    && last_scroll_region_right < screen_w
-                    && last_scroll_region_right % ts != 0
-                {
-                    Some(last_scroll_region_right / ts)
-                } else {
-                    None
-                };
-                let boundary_top_row = if use_static_split
-                    && last_scroll_region_top > 0
-                    && last_scroll_region_top % ts != 0
-                {
-                    Some(last_scroll_region_top / ts)
-                } else {
-                    None
-                };
-                let boundary_bottom_row = if use_static_split
-                    && last_scroll_region_bottom < screen_h
-                    && last_scroll_region_bottom % ts != 0
-                {
-                    Some(last_scroll_region_bottom / ts)
-                } else {
-                    None
-                };
-                let tiles = emitter.emit_static_tiles(
-                    &rgba,
-                    stride,
-                    &static_dirty,
-                    &grid,
-                    boundary_col,
-                    boundary_top_row,
-                    boundary_bottom_row,
-                    editable_qoi_tile_bounds,
-                );
-                let draw_mode_on =
-                    bpane_protocol::TileMessage::TileDrawMode { apply_offset: true }.to_frame();
-                let mut out = Vec::with_capacity(tiles.len() + 2);
-                out.push(draw_mode_off);
-                out.extend(tiles);
-                out.push(draw_mode_on);
-                out
-            } else {
-                Vec::new()
-            };
-            grid.clear_dirty();
-
-            let static_tile_count = static_frames.len().saturating_sub(2); // minus DrawMode on/off
-            let static_bytes: usize = static_frames.iter().map(|f| f.payload.len()).sum();
-
-            let s = &result.stats;
-            if s.fills > 0 || s.qoi_tiles > 0 || s.cache_hits > 0 || scroll_residual_ratio.is_some()
-            {
-                let scroll_saved_rate_total = if scroll_potential_tiles_total > 0 {
-                    Some(scroll_saved_tiles_total as f32 / scroll_potential_tiles_total as f32)
-                } else {
-                    None
-                };
-                tracing::trace!(
-                    skipped = s.skipped,
-                    fills = s.fills,
-                    cache_hits = s.cache_hits,
-                    qoi = s.qoi_tiles,
-                    video_changed = s.video_tiles,
-                    video_latched = latched_video_tiles.len(),
-                    cdp_motion_tiles,
-                    cdp_hint_raw = cdp_hint_region_raw.is_some(),
-                    cdp_video_hint = cdp_video_region_hint.is_some(),
-                    editable_qoi = editable_qoi_tile_bounds.is_some(),
-                    key_input_qoi_boost,
-                    click_armed = cdp_click_armed,
-                    scroll_suppress = scroll_cooldown_frames,
-                    scroll_residual_ratio = scroll_residual_ratio.map(|r| format!("{:.2}", r)),
-                    scroll_residual_full = scroll_residual_fallback_full,
-                    scroll_residual_tiles = scroll_residual_tiles_frame,
-                    scroll_potential_tiles = scroll_potential_tiles_frame,
-                    scroll_saved_tiles = scroll_saved_tiles_frame,
-                    scroll_saved_ratio = scroll_saved_ratio_frame.map(|r| format!("{:.2}", r)),
-                    scroll_emit_ratio = scroll_emit_ratio_frame.map(|r| format!("{:.2}", r)),
-                    scroll_batches_total = scroll_residual_batches_total,
-                    scroll_full_fallbacks_total = scroll_residual_fallback_full_total,
-                    scroll_potential_tiles_total,
-                    scroll_residual_tiles_total,
-                    scroll_saved_tiles_total,
-                    scroll_saved_rate_total = scroll_saved_rate_total.map(|r| format!("{:.2}", r)),
-                    scroll_thin_mode = scroll_thin_mode_frame,
-                    scroll_thin_repair = scroll_thin_repair_frame,
-                    scroll_thin_active = scroll_thin_mode_active,
-                    scroll_thin_batches_total,
-                    scroll_thin_repairs_total,
-                    qoi_kb = s.qoi_bytes / 1024,
-                    static_tiles = static_tile_count,
-                    static_bytes,
-                    content_dirty = content_dirty.len(),
-                    static_dirty_count = static_dirty.len(),
-                    use_static_split,
-                    "tile frame"
-                );
-            }
-
-            // When there are video tiles, let H.264 through for the video
-            // region. Otherwise tiles handle the full screen.
-            let cdp_has_video = cdp_video_region_hint.is_some() && scroll_cooldown_frames == 0;
-            let _cdp_has_motion =
-                cdp_has_video && cdp_motion_tiles >= MIN_CHANGED_VIDEO_TILES_FOR_H264;
-            let has_video = cdp_has_video;
-            let desired_h264 = match h264_mode_for_tile {
-                H264Mode::Always => true,
-                H264Mode::VideoTiles => has_video,
-                H264Mode::Off => false,
-            };
-
-            let next_capture_region = if matches!(h264_mode_for_tile, H264Mode::VideoTiles) {
-                cdp_has_video.then_some(cdp_video_region_hint).flatten()
-            } else {
-                None
-            };
-            let committed_capture_region =
-                region_committer.commit(next_capture_region, REGION_RECONFIG_STABLE_FRAMES);
-
-            if desired_h264 && committed_capture_region != region_committer.active {
-                if region_committer.should_reconfig(
-                    committed_capture_region,
-                    now,
-                    REGION_RECONFIG_MIN_INTERVAL_MS,
-                ) {
-                    region_committer.apply_reconfig(committed_capture_region, now);
-                    let _ = cmd_tx_for_tile.send(capture::ffmpeg::PipelineCmd::SetRegion(
-                        committed_capture_region,
-                    ));
-                }
-            } else if !desired_h264 {
-                region_committer.active = committed_capture_region;
-            }
-
-            let next_tile_info = if matches!(h264_mode_for_tile, H264Mode::VideoTiles) {
-                region_committer.active.map(|region| VideoTileInfo {
-                    tile_x: region.x as u16,
-                    tile_y: region.y as u16,
-                    tile_w: region.w as u16,
-                    tile_h: region.h as u16,
-                    screen_w,
-                    screen_h,
-                })
-            } else {
-                None
-            };
-            {
-                let mut guard = match video_tile_info_for_tile.lock() {
-                    Ok(g) => g,
-                    Err(poisoned) => poisoned.into_inner(),
-                };
-                *guard = next_tile_info;
-            }
-
-            let mut effective_h264 = desired_h264;
-            if h264_enabled
-                && !desired_h264
-                && now.duration_since(last_h264_toggle_at)
-                    < std::time::Duration::from_millis(H264_MIN_ON_DURATION_MS)
-            {
-                effective_h264 = true;
-            }
-            if effective_h264 != h264_enabled {
-                h264_enabled = effective_h264;
-                last_h264_toggle_at = now;
-                let _ =
-                    cmd_tx_for_tile.send(capture::ffmpeg::PipelineCmd::SetEnabled(effective_h264));
-            }
-            let tiles_cover_screen = if matches!(h264_mode_for_tile, H264Mode::Off) {
-                true
-            } else {
-                !effective_h264 || region_committer.active.is_none()
-            };
-            tiles_active_for_tile.store(tiles_cover_screen, std::sync::atomic::Ordering::Relaxed);
-
-            // Send all tile data (content + static) BEFORE BatchEnd so the
-            // client processes everything in a single batch.  Previously,
-            // static tiles were sent after BatchEnd, causing them to be
-            // deferred to the next batch — leaving the buffer row (header/
-            // content seam) black for one frame during scroll.
-            //
-            // Order: content tiles (sans BatchEnd) → static tiles → BatchEnd.
-            let mut content_frames = result.tile_frames;
-            let batch_end_frame = content_frames.pop(); // always BatchEnd
-            for frame in content_frames {
-                if tile_tx.blocking_send(frame).is_err() {
-                    return;
-                }
-            }
-            for frame in static_frames {
-                if tile_tx.blocking_send(frame).is_err() {
-                    return;
-                }
-            }
-            if let Some(be) = batch_end_frame {
-                if tile_tx.blocking_send(be).is_err() {
-                    return;
-                }
-            }
-
-            if let Some(dt) = damage.as_mut() {
-                dt.reset();
-            }
-
-            // Keep current frame as previous without cloning.
-            prev_frame = Some(rgba);
-        }
     });
+
 
     // Forward tile frames to the gateway
     let tile_gateway = to_gateway.clone();
