+++ code/apps/bpane-host/src/tile_loop/tests.rs
@@ -1,0 +1,2 @@
+// Integration tests for the tile capture thread will be added
+// when the full session test harness is available.
